Updates captions for the three regions
