--- support-frontend/assets/pages/showcase/components/regionalContent.js
@@ -1,33 +1,0 @@
-
-export const regionalContent = {
-  ukContent: {
-    title: 'Cambridge Analytica',
-    caption: 'A year-long investigation in which we revealed how the data analytics firm Cambridge Analytica that was behind Trump\'s 2016 campaign and played a role in Brexit, had used the data harvested from 87 million Facebook users without their consent. This reporting led to a public apology from Facebook\'s Mark Zuckerberg who was forced to testify before Congress.',
-    images: {
-      first: 'showcaseChrisSquare',
-      second: 'showcaseZuckGlass',
-      third: 'showcaseNix',
-      fourth: 'showcaseBrit',
-    },
-  },
-  auContent: {
-    title: 'Cambridge Analytica',
-    caption: 'A year-long investigation in which we revealed how the data analytics firm Cambridge Analytica that was behind Trump\'s 2016 campaign and played a role in Brexit, had used the data harvested from 87 million Facebook users without their consent. This reporting led to a public apology from Facebook\'s Mark Zuckerberg who was forced to testify before Congress.',
-    images: {
-      first: 'newsroom',
-      second: 'showcaseZuckGlass',
-      third: 'protestorsWide',
-      fourth: 'showcaseBrit',
-    },
-  },
-  usContent: {
-    title: 'Cambridge Analytica',
-    caption: 'A year-long investigation in which we revealed how the data analytics firm Cambridge Analytica that was behind Trump\'s 2016 campaign and played a role in Brexit, had used the data harvested from 87 million Facebook users without their consent. This reporting led to a public apology from Facebook\'s Mark Zuckerberg who was forced to testify before Congress.',
-    images: {
-      first: 'showcaseChrisSquare',
-      second: 'showcaseZuckGlass',
-      third: 'showcaseNix',
-      fourth: 'showcaseBrit',
-    },
-  },
-};
