+++ support-frontend/assets/pages/showcase/components/regionalContent.jsx
@@ -1,0 +1,47 @@
+import React from 'react';
+
+export const regionalContent = {
+  ukContent: {
+    title: 'Windrush',
+    caption: (
+      <span>
+        In 2018, we revealed that the British Home Office had wrongly detained and threatened with deportation&#32;
+        hundreds of UK residents. Among their number were members of the Windrush Generation &mdash; named after&#32;
+        the <span className="showcase-hero__caption--italics">Empire Windrush</span>, the ship that brought one of the&#32;
+        first groups of West Indian migrants to the UK. The scandal prompted a wider debate about British&#32;
+        immigration policy.
+      </span>),
+    images: {
+      first: 'showcaseChrisSquare',
+      second: 'showcaseZuckGlass',
+      third: 'showcaseNix',
+      fourth: 'showcaseBrit',
+    },
+  },
+  auContent: {
+    title: 'Climate crisis',
+    caption: (
+      <span>
+        <span className="showcase-hero__caption--italics">Guardian Australia</span> gives climate crisis&#32;
+        reporting the prominence it deserves, stories that often go unreported in the mainstream media.&#32;
+        At this critical time for our planet, we are determined to inform readers about threats,&#32;
+        consequences and solutions based on scientific facts, not political prejudice or business interests.
+      </span>),
+    images: {
+      first: 'newsroom',
+      second: 'showcaseZuckGlass',
+      third: 'protestorsWide',
+      fourth: 'showcaseBrit',
+    },
+  },
+  usContent: {
+    title: 'Cambridge Analytica',
+    caption: 'A year-long investigation in which we revealed that 50 million Facebook profiles had been harvested without their consent by data analytics firm Cambridge Analytica, which worked with Trump\'s election team and the Brexit campaign.',
+    images: {
+      first: 'showcaseChrisSquare',
+      second: 'showcaseZuckGlass',
+      third: 'showcaseNix',
+      fourth: 'showcaseBrit',
+    },
+  },
+};
